Added JWT after authentification
