+++ package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "body-parser": "^1.15.1",
     "express": "^4.13.4",
+    "json-web-token": "^2.0.2",
     "method-override": "^2.3.5",
     "mocha": "^2.5.3",
     "mocha-istanbul": "^0.2.0",
