+++ src/Connection.html
@@ -36,6 +36,24 @@
 <script>
     var dbListener = new DbListener(true, false);
     var param = {speed: 5, step: 100};
+    function resizePaper(){
+        drawing.dimension(document.body.clientWidth,document.documentElement.clientHeight);//attr("preserveAspectRatio", "xMinYMin meet") ;
+        drawings.glass.dimension(drawing.width,drawing.height).position(drawing.width/2,drawing.height/2);
+
+        switch(drawing.currentPageDisplayed){
+            case "FormationsManager":
+                mainManipulator.flush();
+                formationsManager.clippingManipulator.last.remove(formationsManager.panel.component);
+                formationsManager.header.display();
+                formationsManager.display();
+                break;
+        }
+    }
+
+    setTimeout(function(){
+        window.onresize = resizePaper;
+    },200);
+
     connexion();
 </script>
 </body>
