+++ src/Domain.js
@@ -1266,6 +1266,50 @@ ConnectionManager = function () {
 
     self.connectionButtonLabel = "Connexion";
     self.tabForm =[];
+
+    let listFormations = function() {
+        let callback = function (data) {
+            let myFormations = JSON.parse(data).myCollection;
+            console.log(myFormations);
+            formationsManager = new FormationsManager(myFormations);
+            formationsManager.display();
+        };
+        dbListener.httpGetAsync("/getAllFormationsNames", callback);
+    }
+
+    self.connectionButtonHandler = function(){
+        var emptyAreas = self.tabForm.filter(field=> field.label === "");
+        emptyAreas.forEach(function(emptyArea){
+            emptyArea.cadre.color(myColors.white, 3, myColors.red);
+        });
+        if (emptyAreas.length>0){
+            var message = autoAdjustText(EMPTYFIELDERROR, 0, 0, drawing.width, self.h, 20, null, self.connectionButtonManipulator, 3);
+            message.text.color(myColors.red).position(0, - self.connectionButton.cadre.height+MARGIN);
+            svg.timeout(function(){
+                self.connectionButtonManipulator.ordonator.unset(3);
+                emptyAreas.forEach(function(emptyArea){
+                    emptyArea.cadre.color(myColors.white, 1, myColors.black);
+                });
+            },5000);
+        } else {
+            let callback = function(data) {
+                token = data ? JSON.parse(data).token : '';
+                if(token){
+                    listFormations();
+                }else {
+                    var message = autoAdjustText("Adresse et/ou mot de passe invalide(s)", 0, 0, drawing.width, self.h, 20, null, self.connectionButtonManipulator, 3);
+                    message.text.color(myColors.red).position(0, - self.connectionButton.cadre.height+MARGIN);
+                    svg.timeout(function(){
+                        self.connectionButtonManipulator.ordonator.unset(3);
+
+                    },5000);
+                }
+            };
+            dbListener.httpPostAsync("/connectUser/" ,{mailAddress: self.mailAddressField.label,password:self.passwordField.labelSecret}, callback);
+
+
+        }
+    };
 };
 
 ////////////////// end of QuizzManager.js //////////////////////////
